+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 
 
-## [v0.7.8] - 
+## [v0.7.8] -
 
 Changes from v0.7.8-rc1 to v0.7.8.
 
@@ -186,5 +186,9 @@ Changes from v0.7.7 to v0.7.8-rc1.
 [See full release details][release-v0.7.8-rc1]
 
 [release-v0.7.8-rc1]: https://github.com/danny-avila/LibreChat/releases/tag/v0.7.8-rc1
+
+## [v0.2.0] - Together AI integration
+
+- feat: add Together AI provider for chat, embeddings, and code interpreter
 
 ---